replaced uint16_t and uint8_t for normal int, and fixed errors after change

+++ src/environment.h
@@ -6,13 +6,13 @@
 #include "utils.h"
 
 struct Environment {
-  uint8_t *memory;
+  int *memory;
   unsigned int PC;
   unsigned int data_offset; // To prevent overwriting code, and to signal end of code
-  uint16_t *registers;
+  int *registers;
 };
 
-void add_instruction(uint8_t *parsed_instruction, struct Environment *env);
+void add_instruction(int *parsed_instruction, struct Environment *env);
 void execute_instruction(struct Environment *env);
 
 #endif
